requirements.txt
Flask==2.3.3
firebase-admin==6.2.0
Werkzeug==2.3.7
Flask-WTF==1.1.1
google-generativeai==0.3.2
python-dotenv==1.0.0
transformers==4.36.0
torch==2.4.1
requests==2.31.0
opencv-python==4.8.1.78
Pillow==10.0.1
mediapipe==0.10.21
authlib==1.2.1
cryptography==41.0.7
